improvement: grocery aisles

+++ Data/Migrations/CoreContextModelSnapshot.cs
@@ -186,8 +186,8 @@ namespace Data.Migrations
                     b.Property<long>("Allergens")
                         .HasColumnType("bigint");
 
-                    b.Property<int>("Category")
-                        .HasColumnType("integer");
+                    b.Property<long>("Category")
+                        .HasColumnType("bigint");
 
                     b.Property<int>("DefaultMeasure")
                         .HasColumnType("integer");
@@ -228,8 +228,8 @@ namespace Data.Migrations
                     b.Property<string>("Notes")
                         .HasColumnType("text");
 
-                    b.Property<int>("Section")
-                        .HasColumnType("integer");
+                    b.Property<long>("Section")
+                        .HasColumnType("bigint");
 
                     b.Property<bool>("SkipShoppingList")
                         .HasColumnType("boolean");
